huggingface-deploy/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic[email]
pydantic-settings
httpx
qdrant-client
google-generativeai
requests
numpy
beautifulsoup4
PyJWT
